New: Options to run from `git` and use the browser extension bundle
Facilitate testing by including an option to create images
from the `hint` repository on GitHub. Default behavior
for production continues to be to use the `hint` package
published on `npm`.

Also add an option to run the bundle generated by
`extension-browser` on live sites using `puppeteer`.
Default behavior for production continues to be to
use the `hint` package directly.

Co-authored-by: Jesus David Garcia Gomez <jdgarcia@outlook.com>